Preserve relationship data items when model is missing
When the model is unable to be built from a normalized reduxObject, the current functionality is to do nothing. What this means is that after denormalizing those objects, it appears that the relationship is null or empty.

It seems reasonable to expect that the relationship links would remain in the denormalized object even if the target of the actual model object can't be found in the normalized store. This issue is addressed by falling back to the relationship data object when `buildModel` returns null.