Fix duplicate model bug
This pull request fixes a bug where a model is both in the main data
and in the include list. If the model from the include list is seen
first it causes jsona to use that model for the main data as well.

Models from the main data shouldn't be fetched from the cache and always
be build correctly.