Fix duplicate model bug
This pull request fixes a bug where a model is both in the main response
and in the include list. If the model from the include list is seen
first it causes jsona to use that model for the main response as well,
even if it has less properties or relations.

In theory this shouldn't happen but some jsonapi serializers such as the
Go one don't deduplicate models and might include them in both places.